add optional lock file during camera init
While the device_num parameter provides a crude way to synchronize
bringup using a shared memory segment and a semaphore, this mechanism
is brittle, and only designed to work once, as it does not handle
shutdown and restart of only one of the cameras. It also requires
knowing the number of cameras a priori in a parameter and requires
more coordination than a simple global advisory lock file. The
advisory lock is simpler, and more robust, as on any process exit it
will be freed by the operating system.