Correct Astra Mino Pro SXGA intrinsics
Correct IR camera_info for 1280x1024 (SXGA) on Astra Mini Pro.
From inspection, the reported camera intrinsics in 1280x1024 is
identical to 1280x960, and 1280x960 intrinsics are an exact 2x of VGA
(640x480). However, the 1280x960 image centers perfectly on the
1280x1024 image, meaning the reported central y point for SXGA is 32
pixels too far up in the image. Fix this by adding 32.