do not fail ROS service for toggle sensor
Do not fail the ROS service when a sensor is already on or off when
the toggle service is called. Instead, return true (so the service
call succeeds). Whether or not something was actually done is already
retuned in the result. Failing the service causes an exception in the
client.