shutdown if device is disconnected
Instead of attempting to reconnect in the driver, simply shutdown if
the active device is disconnected. This lets the system restart the
service, if desired, instead of attempting to reinitialize in the
driver.

Note from Howard: During rebase from upstream 1.2.2 to 1.3.8, I changed
this commit. The original version deleted the device reset and added a
shutdown of the ROS node. I retained the device reset and simply added
the shutdown of the ROS node. This would appear to be a cleaner way to
exit. Making a note of the change here in case we find that it causes
problems.