badger: provide dyncfg for disabling emitter
Provide a legacy (backwards-compatibile) interface via dynamic
reconfigure to disable the IR emitter (laser). It would take a bit of
work to hunt down the various places that either disable the emitter
or need to know its state and figure out how to rearchitect them. In
at least one place, the code is relying on the dynamic reconfigure
parameter update topic to get changes to the emitter state. One way
to rearchitect would be to use resource manager to send state changes
and have it manage tracking the state of the emitter.