badger: add optional lock file during camera init
While the driver now contains a inter-process shared-mutex that it
locks during initialization, it does not protect context creation or
device enumeration (which has been observed to issue USB traffic to
other devices causing problems with other camera's initialization)
nor does it synchronize with the old ros astra driver. Badger has
already added this same flock mechanism to the old ros astra driver,
so using it in the orbbec driver synchronizes both drivers properly
during initialization/startup. Also, using flock is more robust as
the operating system will automatically unlock the file when the file
is closed on exit, meaning there can be no path where the lock is
left in a locked state on failure.