Fix setting IR exposure
Fix the name of the expected exposure param to drop the trailing
underscore. I do not believe this was intended as all the launch
files for the cameras refer to it as 'ir_exposure'.

Also move the setting of the IR auto exposure to before the
setting of the IR exposure. I found that this was re-setting
the exposure to the default value of 3000.