Abort list_devices if systemd services are running
Previously, attempting to run the list_devices_node to
get all connected Orbbec cameras while the systemd services
were running was impossible. After switching to v4l as the
UVC backed this is no longer true. Running the list_devices
node while the systemd services are running though breaks
the v4l2 udev port and requires the truck to be rebooted.

Check all units for ones that start with ros.depthcam, and
do not attempt to query the devices if any are active.